add bundle-jessie-factory

+++ lib/functions.sh
@@ -125,8 +125,13 @@ export ANSIBLE_HOST_KEY_CHECKING=False
     ansible-playbook ./$2/ansible/bootstrap.yml -i host_inventory
     ansible-playbook purge_image_fe.yml -i host_inventory
  fi
+ if [ "$?" != "0" ]; then
+  echo "Ansible failed"
+  exit 1
+fi
 
 rm -f host_inventory
+
 
 }
